refactor: move reservationDTOs to new package
