+++ rentplace/src/main/java/kattsyn/dev/rentplace/controllers/ReservationController.java
@@ -9,7 +9,8 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import kattsyn.dev.rentplace.dtos.*;
+import kattsyn.dev.rentplace.dtos.reservations.ReservationCreateEditDTO;
+import kattsyn.dev.rentplace.dtos.reservations.ReservationDTO;
 import kattsyn.dev.rentplace.services.ReservationService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
