+++ rentplace/src/main/java/kattsyn/dev/rentplace/dtos/reservations/ReservationCreateEditDTO.java
@@ -1,4 +1,4 @@
-package kattsyn.dev.rentplace.dtos;
+package kattsyn.dev.rentplace.dtos.reservations;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
