+++ rentplace/src/main/java/kattsyn/dev/rentplace/services/ReservationService.java
@@ -1,7 +1,7 @@
 package kattsyn.dev.rentplace.services;
 
-import kattsyn.dev.rentplace.dtos.ReservationCreateEditDTO;
-import kattsyn.dev.rentplace.dtos.ReservationDTO;
+import kattsyn.dev.rentplace.dtos.reservations.ReservationCreateEditDTO;
+import kattsyn.dev.rentplace.dtos.reservations.ReservationDTO;
 import kattsyn.dev.rentplace.entities.Reservation;
 import org.springframework.stereotype.Service;
